WebScrape for DID electrical website
Utilizing web scraping techniques for the purpose of acquiring information on a TV, the extraction process retrieves both the title and price of the product. The obtained data is organized and input into a CSV file. A timer has been implemented to execute the code at intervals of one minute, to append additional information to the existing CSV file.